rattler_conda_types

Module match_spec

Query language for conda packages.

Modules

condition

Experimental conditionals for match specs.

matcher

Match a given string either by exact match, glob or regex

package_name_matcher

Match package names either by exact match, glob or regex

parse

Parse a match spec from a string

Structs

MatchSpec

A MatchSpec is, fundamentally, a query language for conda packages. Any of the fields that comprise a crate::PackageRecord can be used to compose a MatchSpec.

NamelessMatchSpec

Similar to a MatchSpec but does not include the package name. This is useful in places where the package name is already known (e.g. foo = "3.4.1 *cuda")

Enums

MatchSpecUrlError

Errors that can occur when converting a URL to a MatchSpec

Traits

Matches

A trait that defines the behavior of matching a spec against a record.