rattler_conda_types

Module menuinst

Define types that can be serialized into a PrefixRecord to track menu entries installed into the system.

Structs

LinuxRegisteredMimeFile

Registered MIME file on the system

LinuxTracker

Tracker for Linux installations

MacOsTracker

Tracker for macOS installations

WindowsFileExtension

File extension that was installed

WindowsTerminalProfile

Terminal profile that was installed

WindowsTracker

Tracker for Windows installations

WindowsUrlProtocol

URL protocol that was installed

Enums

MenuMode

Menu mode that was used to install the menu entries

Tracker

Tracker for menu entries installed into the system