rattler_conda_types

Module minimal_prefix_record

Minimal prefix record reading for fast environment change detection.

This module provides functionality to read only the minimal metadata needed from conda-meta JSON files to determine if packages have changed, avoiding the expensive parsing of file lists and other large fields of PrefixRecord.

The most interesting part of this file is custom parser written using nom. It uses PrefixRecord structure to parse as little data as possible, also uses byte parsing on memory map, so we don’t read file entirely. See documentation of MinimalPrefixRecord::from_path for more information.

Structs

MinimalPrefixRecord

A minimal version of PrefixRecord that only contains fields needed for transaction computation. This is much faster to parse than the full PrefixRecord.

Constants

MMAP_HEURISTIC_THRESHOLD

We want to use memory map if file size is larger than 1MB to avoid doing many small reads, as this can be inefficient.

Traits

MinimalPrefixCollection

Extension trait for PrefixRecord to support sparse collection

Functions

collect_minimal_prefix_records

Collect minimal prefix records from a prefix directory. This is much faster than collecting full PrefixRecords when you only need to check if packages have changed.