rattler_conda_types::minimal_prefix_record

Trait MinimalPrefixCollection

pub trait MinimalPrefixCollection {
    // Required method
    fn collect_minimal_from_prefix(
        prefix: &Path,
    ) -> Result<Vec<MinimalPrefixRecord>, Error>;
}

Extension trait for PrefixRecord to support sparse collection

Required Methods

fn collect_minimal_from_prefix( prefix: &Path, ) -> Result<Vec<MinimalPrefixRecord>, Error>

Collect only the minimal fields needed for transaction computation. Fall back to full parsing if more fields needed!

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.