rattler_conda_types

Module package

Contains models of files that are found in the info/ directory of a package.

Structs

AboutJson

The about.json file contains metadata about the package

ArchiveIdentifier

A package in the conda ecosystem consists of a name, version and build_string. This can be used to uniquely identify a package in a subdirectory.

CondaArchiveIdentifier

A conda package archive identifier contains the name, version, build_string and archive_type of a conda package archive. This information can be derived from the filename of a conda package archive using the CondaArchiveIdentifier::try_from_filename and CondaArchiveIdentifier::try_from_url functions.

DistArchiveIdentifier

A DistArchiveIdentifier is similar to a CondaArchiveIdentifier but represents a package in a format that is not a conda native archive. A DistArchiveIdentifier can be used to uniquely identify an archive, but it does not necessarily represent the correct filename of the archive.

EntryPoint

A struct for a single Python entry point. An entry point is a command that runs a function in a Python module. For example, the entry point jlpm = jupyterlab.jlpmapp:main will run the main function in the jupyterlab.jlpmapp module when the command jlpm is run on the command line. The main usage for entry points is in noarch: python packages.

Files

Representation of the info/files file in older package archives.

HasPrefix

Representation of the info/has_prefix file in older package archives.

HasPrefixEntry

Representation of an entry in info/has_prefix.

IndexJson

A representation of the index.json file found in package archives.

LinkJson

A representation of the link.json file found in noarch package archives.

NoLink

Representation of the info/no_link file in older package archives. This file contains a list of all files that should not be “linked” (i.e. hard linked) but copied.

NoSoftlink

Representation of the info/no_softlink file in older package archives. This file contains a list of all files that should not be “softlink-ed”.

PackageMetadata

This is metadata about the package version that is contained in .conda packages only, in the outer zip archive.

PathsEntry

A single entry in the paths.json file.

PathsJson

A representation of the paths.json file found in package archives.

PrefixPlaceholder

Description off a placeholder text found in a file that must be replaced when installing the file into the prefix.

PythonEntryPoints

Describes python noarch specific entry points

RunExportsJson

A representation of the run_exports.json file found in package archives.

Enums

CondaArchiveType

Describes the type of a conda package archive.

DistArchiveType

Describes any type of distributable package archive (conda packages or wheels).

FileMode

The file mode of the entry

NoArchLinks

Links for specific types of noarch packages.

PathType

The path type of the path entry

WheelArchiveType

Describes the type of a wheel package archive.

Traits

PackageFile

A trait implemented for structs that represent specific files in a Conda archive.