rattler_conda_types::package

Struct IndexJson

pub struct IndexJson {

    pub arch: Option<String>,
    pub build: String,
    pub build_number: u64,
    pub constrains: Vec<String>,
    pub depends: Vec<String>,
    pub experimental_extra_depends: BTreeMap<String, Vec<String>>,
    pub features: Option<String>,
    pub license: Option<String>,
    pub license_family: Option<String>,
    pub name: PackageName,
    pub noarch: NoArchType,
    pub platform: Option<String>,
    pub purls: Option<BTreeSet<PackageUrl>>,
    pub python_site_packages_path: Option<String>,
    pub subdir: Option<String>,
    pub timestamp: Option<TimestampMs>,
    pub track_features: Vec<String>,
    pub version: VersionWithSource,

}

A representation of the index.json file found in package archives.

The index.json file contains information about the package build and dependencies of the package. This data makes up the repodata.json file in the repository.

Fields
arch: Option<String>

Optionally, the architecture the package is build for.

build: String

The build string of the package.

build_number: u64

The build number of the package. This is also included in the build string.

constrains: Vec<String>

The package constraints of the package

depends: Vec<String>

The dependencies of the package

experimental_extra_depends: BTreeMap<String, Vec<String>>

Extra dependency groups that can be selected using foobar[extras=["scientific"]] The implementation is specified in this CEP: https://github.com/conda/ceps/pull/111

features: Option<String>

Features are a deprecated way to specify different feature sets for the conda solver. This is not supported anymore and should not be used. Instead, mutex packages should be used to specify mutually exclusive features.

license: Option<String>

Optionally, the license

license_family: Option<String>

Optionally, the license family

name: PackageName

The lowercase name of the package

noarch: NoArchType

If this package is independent of architecture this field specifies in what way. See NoArchType for more information.

platform: Option<String>

Optionally, the OS the package is build for.

purls: Option<BTreeSet<PackageUrl>>

A list of Package URLs identifying this package. See this CEP: https://github.com/conda/ceps/pull/63

python_site_packages_path: Option<String>

Optionally a path within the environment of the site-packages directory. This field is only present for python interpreter packages. This field was introduced with https://github.com/conda/ceps/blob/main/cep-17.md.

subdir: Option<String>

The subdirectory that contains this package

timestamp: Option<TimestampMs>

The timestamp when this package was created

track_features: Vec<String>

Track features are nowadays only used to downweight packages (ie. give them less priority). To that effect, the number of track features is counted (number of commas) and the package is downweighted by the number of track_features.

version: VersionWithSource

The version of the package