rattler_conda_types::package

Trait PackageFile

pub trait PackageFile: Sized {
    // Required methods
    fn package_path() -> &'static Path;
    fn from_str(str: &str) -> Result<Self, Error>;

    // Provided methods
    fn from_slice(slice: &[u8]) -> Result<Self, Error> { ... }
    fn from_reader(reader: impl Read) -> Result<Self, Error> { ... }
    fn from_path(path: impl AsRef<Path>) -> Result<Self, Error> { ... }
    fn from_package_directory(path: impl AsRef<Path>) -> Result<Self, Error> { ... }
}

A trait implemented for structs that represent specific files in a Conda archive.

This trait provides a standardized interface for accessing the contents of known files in a Conda package, such as the index.json (see IndexJson) or about.json (see AboutJson) files. Structs that represent these files should implement this trait in order to ensure that they can be easily accessed and manipulated by other code that expects a consistent interface.

Required Methods

fn package_path() -> &'static Path

Returns the path to the file within the Conda archive.

The path is relative to the root of the archive and include any necessary directories.

fn from_str(str: &str) -> Result<Self, Error>

Parses the object from a string, using a format appropriate for the file type.

For example, if the file is in JSON format, this function parses the JSON string and returns the resulting object. If the file is not in a parse-able format, this function returns an error.

Provided Methods

fn from_slice(slice: &[u8]) -> Result<Self, Error>

Parses the object from a byte slice, using a format appropriate for the file type.

fn from_reader(reader: impl Read) -> Result<Self, Error>

Parses the object from a Read trait object, using a format appropriate for the file type.

For example, if the file is in JSON format, this function reads the data from the Read object, parse the JSON string and return the resulting object. If the file is not in a parse-able format, this function returns an error.

fn from_path(path: impl AsRef<Path>) -> Result<Self, Error>

Parses the object from a file specified by a path, using a format appropriate for the file type.

For example, if the file is in JSON format, this function reads the data from the file at the specified path, parse the JSON string and return the resulting object. If the file is not in a parse-able format or if the file could not read, this function returns an error.

fn from_package_directory(path: impl AsRef<Path>) -> Result<Self, Error>

Parses the object by looking up the appropriate file from the root of the specified Conda archive directory, using a format appropriate for the file type.

For example, if the file is in JSON format, this function reads the appropriate file from the archive, parse the JSON string and return the resulting object. If the file is not in a parse-able format or if the file could not be read, this function returns an error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.