rattler_conda_types

Module prefix_record

Defines the [PrefixRecord] struct.

Structs

Link

A record of a single file that was installed into the prefix

PathsEntry

Information about a single file installed for a package.

PrefixPaths

Information about every file installed with the package.

PrefixRecord

A record of a single package installed within an environment. The struct includes the RepoDataRecord which specifies information about where the original package comes from.

Enums

LinkType

The different link types that are used

PathType

Information about a single file installed for a package.