rattler_conda_types

Struct PackageRecord

pub struct PackageRecord {

    pub arch: Option<String>,
    pub build: String,
    pub build_number: BuildNumber,
    pub constrains: Vec<String>,
    pub depends: Vec<String>,
    pub experimental_extra_depends: BTreeMap<String, Vec<String>>,
    pub features: Option<String>,
    pub legacy_bz2_md5: Option<Md5Hash>,
    pub legacy_bz2_size: Option<u64>,
    pub license: Option<String>,
    pub license_family: Option<String>,
    pub md5: Option<Md5Hash>,
    pub name: PackageName,
    pub noarch: NoArchType,
    pub platform: Option<String>,
    pub purls: Option<BTreeSet<PackageUrl>>,
    pub python_site_packages_path: Option<String>,
    pub run_exports: Option<RunExportsJson>,
    pub sha256: Option<Sha256Hash>,
    pub size: Option<u64>,
    pub subdir: String,
    pub timestamp: Option<TimestampMs>,
    pub track_features: Vec<String>,
    pub version: VersionWithSource,

}

A single record in the Conda repodata. A single record refers to a single binary distribution of a package on a Conda channel.

Fields
arch: Option<String>

Optionally the architecture the package supports. This is almost always the second part of the subdir string. Except for 64 which maps to x86_64 and 32 which maps to x86. This will be None if the package is noarch.

build: String

The build string of the package

build_number: BuildNumber

The build number of the package

constrains: Vec<String>

Additional constraints on packages. constrains are different from depends in that packages specified in depends must be installed next to this package, whereas packages specified in constrains are not required to be installed, but if they are installed they must follow these constraints.

depends: Vec<String>

Specification of packages this package depends on

experimental_extra_depends: BTreeMap<String, Vec<String>>

Specifications of optional or dependencies. These are dependencies that are only required if certain features are enabled or if certain conditions are met.

features: Option<String>

Features are a deprecated way to specify different feature sets for the conda solver. This is not supported anymore and should not be used. Instead, mutex packages should be used to specify mutually exclusive features.

legacy_bz2_md5: Option<Md5Hash>

A deprecated md5 hash

legacy_bz2_size: Option<u64>

A deprecated package archive size.

license: Option<String>

The specific license of the package

license_family: Option<String>

The license family

md5: Option<Md5Hash>

Optionally a MD5 hash of the package archive

name: PackageName

The name of the package

noarch: NoArchType

If this package is independent of architecture this field specifies in what way. See NoArchType for more information.

platform: Option<String>

Optionally the platform the package supports. Note that this does not match the Platform enum, but is only the first part of the platform (e.g. linux, osx, win, …). The subdir field contains the Platform enum.

purls: Option<BTreeSet<PackageUrl>>

Package identifiers of packages that are equivalent to this package but from other ecosystems. starting from 0.23.2, this field became Option<Vec<PackageUrl>>. This was done to support older lockfiles, where we didn’t differentiate between empty purl and missing one. Now, None:: means that the purl is missing, and it will be tried to filled in. So later it can be one of the following: [Some(vec![])] means that the purl is empty and package is not pypi one. [Some([PackageUrl])] means that it is a pypi package. See this CEP: https://github.com/conda/ceps/pull/63

python_site_packages_path: Option<String>

Optionally a path within the environment of the site-packages directory. This field is only present for python interpreter packages. This field was introduced with https://github.com/conda/ceps/blob/main/cep-17.md.

run_exports: Option<RunExportsJson>

Run exports that are specified in the package.

sha256: Option<Sha256Hash>

Optionally a SHA256 hash of the package archive

size: Option<u64>

Optionally the size of the package archive in bytes

subdir: String

The subdirectory where the package can be found

timestamp: Option<TimestampMs>

The date this entry was created.

track_features: Vec<String>

Track features are nowadays only used to downweight packages (ie. give them less priority). To that effect, the package is downweighted by the number of track_features.

version: VersionWithSource

The version of the package

Implementations

impl PackageRecord

pub fn apply_patch(&mut self, patch: &PackageRecordPatch)

Apply a patch to a single package record

impl PackageRecord

pub fn has_run_exports(&self) -> bool

Returns true if package run_exports is some.

impl PackageRecord

pub fn new( name: PackageName, version: impl Into<VersionWithSource>, build: String, ) -> Self

A simple helper method that constructs a PackageRecord with the bare minimum values.

pub fn sort_topologically<T: AsRef<PackageRecord> + Clone>( records: Vec<T>, ) -> Vec<T>

Sorts the records topologically.

This function is deterministic, meaning that it will return the same result regardless of the order of records and of the depends vector inside the records.

Note that this function only works for packages with unique names.

pub fn validate<T: AsRef<PackageRecord>>( records: Vec<T>, ) -> Result<(), Box<ValidatePackageRecordsError>>

Validate that the given package records are valid w.r.t. ‘depends’ and ‘constrains’. This function will return Ok(()) if all records form a valid environment, i.e., all dependencies of each package are satisfied by the other packages in the list. If there is a dependency that is not satisfied, this function will return an error.

impl PackageRecord

pub fn from_index_json( index: IndexJson, size: Option<u64>, sha256: Option<Sha256Hash>, md5: Option<Md5Hash>, ) -> Result<PackageRecord, ConvertSubdirError>

Builds a PackageRecord from a IndexJson and optionally a size, sha256 and md5 hash.