rattler_conda_types::utils

Module url_with_trailing_slash

This module provides utilities for handling and normalizing URLs that are required to end with a trailing slash. It ensures consistency in URL formatting for downstream processing or matching logic.

Structs

UrlWithTrailingSlash

A URL that always has a trailing slash. A trailing slash in a URL has significance but users often forget to add it. This type is used to normalize the use of the URL.