rattler_conda_types::utils::url_with_trailing_slash

Struct UrlWithTrailingSlash

pub struct UrlWithTrailingSlash(/* private fields */);

A URL that always has a trailing slash. A trailing slash in a URL has significance but users often forget to add it. This type is used to normalize the use of the URL.