rattler_conda_types

Struct Version

pub struct Version { /* private fields */ }

This class implements an order relation between version strings. Version strings can contain the usual alphanumeric characters (A-Za-z0-9), separated into segments by dots and underscores. Empty segments (i.e. two consecutive dots, a leading/trailing underscore) are not permitted. An optional epoch number - an integer followed by ‘!’ - can precede the actual version string (this is useful to indicate a change in the versioning scheme itself). Version comparison is case-insensitive.

Rattler supports six types of version strings:

Release versions contain only integers, e.g. ‘1.0’, ‘2.3.5’.

Pre-release versions use additional letters such as ‘a’ or ‘rc’, for example ‘1.0a1’, ‘1.2.beta3’, ‘2.3.5rc3’.

Development versions are indicated by the string ‘dev’, for example ‘1.0dev42’, ‘2.3.5.dev12’.

Post-release versions are indicated by the string ‘post’, for example ‘1.0post1’, ‘2.3.5.post2’.

Tagged versions have a suffix that specifies a particular property of interest, e.g. ‘1.1.parallel’. Tags can be added to any of the preceding four types. As far as sorting is concerned, tags are treated like strings in pre-release versions.

An optional local version string separated by ‘+’ can be appended to the main (upstream) version string. It is only considered in comparisons when the main versions are equal, but otherwise handled in exactly the same manner.

To obtain a predictable version ordering, it is crucial to keep the version number scheme of a given package consistent over time.

Specifically,

version strings should always have the same number of components (except for an optional tag suffix or local version string),

letters/strings indicating non-release versions should always occur at the same position.

Before comparison, version strings are parsed as follows:

They are first split into epoch, version number, and local version number at ‘!’ and ‘+’ respectively. If there is no ‘!’, the epoch is set to 0. If there is no ‘+’, the local version is empty.

The version part is then split into components at ‘.’ and ‘_’.

Each component is split again into runs of numerals and non-numerals

Sub-components containing only numerals are converted to integers.

Strings are converted to lower case, with special treatment for ‘dev’ and ‘post’.

When a component starts with a letter, the fill value 0 is inserted to keep numbers and strings in phase, resulting in ‘1.1.a1’ == 1.1.0a1’.

The same is repeated for the local version part.

Examples:

1.2g.beta15.rc => [[0], [1], [2, 'g'], [0, 'beta', 15], [0, 'rc']] 1!2.15.1_ALPHA => [[1], [2], [15], [1, '_alpha']]

The resulting lists are compared lexicographically, where the following rules are applied to each pair of corresponding sub-components:

integers are compared numerically

strings are compared lexicographically, case-insensitive

strings are smaller than integers, except

‘dev’ versions are smaller than all corresponding versions of other types

‘post’ versions are greater than all corresponding versions of other types

if a sub-component has no correspondent, the missing correspondent is treated as integer 0 to ensure ‘1.1’ == ‘1.1.0’.

The resulting order is:

       0.4
     < 0.4.0
     < 0.4.1.rc
    == 0.4.1.RC   # case-insensitive comparison
     < 0.4.1
     < 0.5a1
     < 0.5b3
     < 0.5C1      # case-insensitive comparison
     < 0.5
     < 0.9.6
     < 0.960923
     < 1.0
     < 1.1dev1    # special case 'dev'
     < 1.1_       # appended underscore is special case for openssl-like versions
     < 1.1a1
     < 1.1.0dev1  # special case 'dev'
    == 1.1.dev1   # 0 is inserted before string
     < 1.1.a1
     < 1.1.0rc1
     < 1.1.0
    == 1.1
     < 1.1.0post1 # special case 'post'
    == 1.1.post1  # 0 is inserted before string
     < 1.1post1   # special case 'post'
     < 1996.07.12
     < 1!0.4.1    # epoch increased
     < 1!3.1.1.6
     < 2!0.4.1    # epoch increased again

Some packages (most notably openssl) have incompatible version conventions. In particular, openssl interprets letters as version counters rather than pre-release identifiers. For openssl, the relation

1.0.1 < 1.0.1a => False # should be true for openssl

holds, whereas conda packages use the opposite ordering. You can work-around this problem by appending an underscore to plain version numbers:

1.0.1_ < 1.0.1a => True # ensure correct ordering for openssl

Implementations

impl Version

pub fn with_alpha(&self) -> Cow<'_, Self>

Add alpha specifier to the end of the version when the last element does not contain an iden component.

For example, 1.0.0 will become 1.0.0.0a0. If the last version element contains a character, it’s not modified (e.g. 1.0.0a will remain 1.0.0a).

pub fn remove_local(&self) -> Cow<'_, Self>

Remove the local segment from the version if it exists. Returns a new version without the local segment.

For example, 1.0.0+3.4 will become 1.0.0.

pub fn bump(&self, bump_type: VersionBumpType) -> Result<Self, VersionBumpError>

Returns a new version after bumping it according to the specified bump type. Note: if a version ends with a character, the next bigger version will use a as the character. For example: 1.1l -> 1.2a, but also 1.1.0alpha -> 1.1.1a.

impl Version

pub fn major(major: u64) -> Version

Constructs a version with just a major component and no other components, e.g. “1”.

pub fn has_epoch(&self) -> bool

Returns true if this version has an epoch.

pub fn has_local(&self) -> bool

Returns true if this version has a local version defined

pub fn epoch(&self) -> u64

Returns the epoch part of the version. If the version did not specify an epoch 0 is returned.

pub fn epoch_opt(&self) -> Option<u64>

Returns the epoch part of the version or None if the version did not specify an epoch.

pub fn segments( &self, ) -> impl DoubleEndedIterator<Item = SegmentIter<'_>> + ExactSizeIterator + '_

Returns the individual segments of the version.

pub fn local_segments( &self, ) -> impl DoubleEndedIterator<Item = SegmentIter<'_>> + ExactSizeIterator + '_

Returns the segments that belong the local part of the version.

The local part of a a version is the part behind the (optional) +. E.g.:

1.2+3.2.1-alpha0
    ^^^^^^^^^^^^ This is the local part of the version

pub fn as_major_minor(&self) -> Option<(u64, u64)>

Tries to extract the major and minor versions from the version. Returns None if this instance doesn’t appear to contain a major and minor version.

pub fn is_dev(&self) -> bool

Returns true if this is considered a dev version.

If a version has a single component named “dev” it is considered to be a dev version.

pub fn starts_with(&self, other: &Self) -> bool

Check if this version version and local strings start with the same as other.

pub fn compatible_with(&self, other: &Self) -> bool

Returns true if this version is compatible with the given other.

pub fn with_segments( &self, segments: impl RangeBounds<usize>, ) -> Option<Version>

Returns a new version with only the given segments.

Calling this function on a version that looks like 1.3a.4-alpha3 with the range [1..3] will return the version: 3a.4.

pub fn pop_segments(&self, n: usize) -> Option<Version>

Pops the specified number of segments from the version. Returns None if the resulting version would become invalid because it no longer contains any segments.

pub fn segment_count(&self) -> usize

Returns the number of segments in the version. Segments are the part of the version separated by dots or dashes.

pub fn strip_local(&self) -> Cow<'_, Version>

Returns either this Version or a new Version where the local version part has been removed.

pub fn extend_to_length( &self, length: usize, ) -> Result<Cow<'_, Version>, VersionExtendError>

Extend the version to the specified length by adding default components (0s). If the version is already longer than the specified length it is returned as is.